requirements.txt
# Core dependencies
pymodbus>=3.0.0
pyserial>=3.5
numpy>=1.24.0

# Input handling
readchar>=4.0.0

# Communication
python-osc>=1.8.0

# GPIO (Raspberry Pi only — install on Pi, skip on dev machines)
gpiozero>=1.6.0

# LED control (Raspberry Pi only — stations 7 and 10)
rpi-ws281x>=5.0.0          # WS2812 addressable LED strip control

# Video and timing
python-vlc>=3.0.0          # RPi HDMI video playback (requires VLC installed)
python-timecode>=1.4.0

# Utilities
PyYAML>=6.0

# --------------------------------------------------------------------------
# Desktop / dev-machine only (not needed on Pi):
#   opencv-python>=4.8.0   — used for local video preview during development
# --------------------------------------------------------------------------
